move proc-log into lib/util where it belongs
