+++ lib/fetcher.js
@@ -9,7 +9,7 @@ const { promisify } = require('util')
 const { basename, dirname } = require('path')
 const rimraf = promisify(require('rimraf'))
 const tar = require('tar')
-const procLog = require('./proc-log.js')
+const procLog = require('./util/proc-log.js')
 const retry = require('promise-retry')
 const fsm = require('fs-minipass')
