+++ lib/util/proc-log.js
